+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Sets `body` to what `arrayBuffer` returns when type not text or JSON.
+
 ## 0.2.1 (Jan 2, 2016)
 - Assigns the `body` property with `Object.defineProperty` to work on Fetch
   implementations that already have `body` as a getter.